Jumpgame program
